feat(S11P21A705-230): 엔딩 로직 추가

See merge request s11-fintech-finance-sub1/S11P21A705!54

+++ backend/src/main/java/com/zzf/backend/domain/animal/dto/AnimalPortfolioResponse.java
@@ -9,9 +9,9 @@ public record AnimalPortfolioResponse(String animalName,
                                       Portfolio portfolio) {
 
     @Builder
-    public record Portfolio(Long depositPercent,
-                            Long savingsPercent,
-                            Long stockPercent,
+    public record Portfolio(Double depositPercent,
+                            Double savingsPercent,
+                            Double stockPercent,
                             String investmentStyle,
                             String ending) {
     }
